Properly convert between binary and list of char representations.
This is needed at least when converting the translations back to lists,
as the generated binaries are in utf-8 format.

I would argue it also makes sense for the input strings, as file
encoding is most likely to be utf-8 (or ascii, which is a proper subset
of unicode) in an application built for internationalization.